Rename field def to mounts
In this newly class-based module, the `def` array is no longer a package global. The list of def directories to mount is now supplied using the `mount` option in new().

As a result, the `def` array is now a purely internal field that can easily be renamed to `mounts` to better reflect how it's used. The reader method had already been named `mounts` in anticipation of this change.