Use country.sii as base for the ATS DLC mount list
When I get a DLC, then downgrade the game to a version before that DLC's release date, init_def() still picks up the DLC because of its filename, which results in inconsistent data.

Instead of only considering the file name, we need to do the same thing the game does: Read the files for a list of supported DLC, and work off of that list. For ATS, that's fairly simple using country.sii in def.scs, although I suspect the game actually does it slightly differently.

Note that states that show up in country.sii may not actually be available as DLC. Such states still need to be skipped. archives() in Archive::SCS::GameDir handles that here.

The same problem probably exists with ETS2, too, but I don't know an easy solution for that.